% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{CSeQTL_run_MatrixEQTL}
\alias{CSeQTL_run_MatrixEQTL}
\title{CSeQTL_run_MatrixEQTL}
\usage{
CSeQTL_run_MatrixEQTL(TREC, RD, XX, SNP, out_cis_fn, cisDist = 1e+06)
}
\arguments{
\item{TREC}{An integer vector of total read counts.}

\item{RD}{A positive numeric vector of log transformed
library sizes per subject.}

\item{XX}{A numeric design matrix of baseline covariates 
including the intercept in the first column and centered 
continuous covariates.}

\item{SNP}{An integer vector of phased genotypes coded 0 (AA),
1 (AB), 2 (BA), 3 (BB), and 5 (NA).}

\item{out_cis_fn}{A full path filename string to store
MatrixEQTL output}

\item{cisDist}{A positive integer for number of SNPs to
include relative to the gene body}
}
\description{
CSeQTL_run_MatrixEQTL
}
