% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSeQTL.R
\name{plot_RHO}
\alias{plot_RHO}
\title{plot_RHO}
\usage{
plot_RHO(RHO, main_plot = "", ...)
}
\arguments{
\item{RHO}{A numeric matrix of cell type proportions. Rows 
correspond to subjects and columns correspond to cell types.}

\item{main_plot}{Character string for the visual's main title.}

\item{...}{Arguments for \code{plot(x,y,bty = "n",...)}.}
}
\description{
plot_RHO
}
