% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{prep_gene_info}
\alias{prep_gene_info}
\title{prep_gene_info}
\usage{
prep_gene_info(work_dir, gtf_gz_fn = NULL)
}
\arguments{
\item{work_dir}{A character string specifying
the working directory to store gene and exon
data.}

\item{gtf_gz_fn}{A character string full path
of the gz compressed gtf file}
}
\value{
Null. No value returned.
}
\description{
Prepare GTF files for gene and 
exon level information
}
