% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSeQTL.R
\name{CSeQTL_run_MatrixEQTL}
\alias{CSeQTL_run_MatrixEQTL}
\title{CSeQTL_run_MatrixEQTL}
\usage{
CSeQTL_run_MatrixEQTL(TREC, RD, XX, SNP, out_cis_fn, cisDist = 1e+06)
}
\arguments{
\item{TREC}{A matrix of integer total read counts, rows are genes
with row labels with syntax "gene_name:chrom:start:end", 
columns are samples with column labels.}

\item{RD}{A positive numeric vector of library sizes per sample.}

\item{XX}{A numeric design matrix of baseline covariates 
including the intercept in the first column and centered 
continuous covariates.}

\item{SNP}{An integer matrix, rows are genomic loci with row labels
such as "chrom:pos", columns correspond to samples with column labels.}

\item{out_cis_fn}{A full path filename string to store
MatrixEQTL output}

\item{cisDist}{A positive integer for number of SNPs to
include relative to the gene body}
}
\description{
CSeQTL_run_MatrixEQTL
}
